requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.7
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.2
fire==0.6.0
firebase-admin==6.5.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
future==1.0.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
msgpack==1.0.8
networkx==3.3
numpy==1.26.4
openai==1.26.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==10.3.0
premailer==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
rarfile==4.2
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
termcolor==2.4.0
tifffile==2024.5.3
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
visualdl==2.5.3
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
